Rashi third

+++ src/Edge.java
@@ -8,7 +8,7 @@ public class Edge implements EdgeData {
     private String info;
     private int tag;
 
-    public Edge(int src, int dest, int weight){
+    public Edge(int src, int dest, double weight){
         this.src = src;
         this.dest = dest;
         this.weight = weight;
